Require cram >= 0.7
I attempted to test with cram 0.5 and 0.6 and getting it to all work
in one codebase was tricky. For the time being, I will just be up-front
that this only works with 0.7 and onwards, but if people ask for
earlier versions to be supported, I can look into this further.